jewelleryai-ms/pyproject.toml
[project]
name = "jewelleryai-ms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pillow>=11.3.0",
    "torch>=2.8.0",
    "transformers>=4.55.0",
]
